Changed type of counters
Changed the variables counter1 and counter2 to int16, so you can put a higher value  for the steps.
Thanks to "tinley" from CCS forum for the suggestion.
//
Cambii en las variables counter1 y counter2 a int16, asi puede darsele un valor mas alto para los pasos.
Gracias a "tinley" del foro de CCS por la sugerencia.

+++ stepper/es/step.c
@@ -113,7 +113,7 @@ void init_stepper(){
 
 void move_stepper(int direction, int16 steps_count){
       int index;
-      int counter1, counter2;
+      int16 counter1, counter2;
       for( counter1=1; counter1<=steps_count; counter1++ ){
          for(counter2=0;counter2<4;counter2++){
             if((mode==HALF_STEP)&&(direction==12)){
